+++ shell_parent_other.go
@@ -1,0 +1,12 @@
+// SPDX-FileType: SOURCE
+// SPDX-FileCopyrightText: 2012 Jesse van den Kieboom
+// SPDX-FileCopyrightText: 2026 Maxim Levchenko (WoozyMasta)
+// SPDX-License-Identifier: BSD-3-Clause
+
+//go:build !windows
+
+package flags
+
+func detectParentShellStyle() RenderStyle {
+	return RenderStyleAuto
+}